Revert "Population Reports Class"
