+++ .github/workflows/main.yml
@@ -1,55 +1,9 @@
-name: A workflow for SemGroup6
+name: A workflow for Group 6 App
 on: push
 
 jobs:
-  UnitTests:
-    name: Unit Tests
-    runs-on: ubuntu-20.04
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v2
-        with:
-          submodules: recursive
-      - name: Set up JDK 11
-        uses: actions/setup-java@v2
-        with:
-          java-version: '11'
-          distribution: 'adopt'
-      - name: Unit Tests
-        run: mvn '-DfailIfNoTests=false' '-Dtest=!com.napier.seprojectgroup6.*Integration*' test
-
-  IntegrationTests:
-    name: Integration Tests
-    runs-on: ubuntu-20.04
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v2
-        with:
-          submodules: recursive
-      - name: Set up JDK 11
-        uses: actions/setup-java@v2
-        with:
-          java-version: '11'
-          distribution: 'adopt'
-      - name: Integration Tests and CodeCov
-        run: |
-          docker build -t database ./db 
-          docker run --name world -dp 33060:3306 database
-          mvn -DfailIfNoTests=false -Dtest=com.napier.seprojectgroup6.*IntegrationTest test          
-          docker stop world
-          docker rm world
-          docker image rm database
-      - name: CodeCov
-        uses: codecov/codecov-action@v4.0.1
-        with:
-          token: ${{ secrets.CODECOV_TOKEN }} # not required for public repos
-          directory: ./target/site/jacoco
-          flags: Integration Tests # optional
-          verbose: true # optional (default = false)
-          slug: KelvinPompey-40718390/sem-project-group6
-
   build:
-    name: Build and Start Using docker-compose
+    name: Basic Functionality Test
     runs-on: ubuntu-20.04
     steps:
       - name: Checkout
@@ -61,7 +15,7 @@ jobs:
         with:
           java-version: '11'
           distribution: 'adopt'
-      - name: Package and Run docker compose
-        run: |
-          mvn package -DskipTests
-          docker-compose up --abort-on-container-exit
+      - name: Build with Maven
+        run: mvn package
+      - name: Run docker compose
+        run: docker-compose up --abort-on-container-exit
