+++ docker-compose.yml
@@ -7,7 +7,5 @@ services:
   # db is is db folder
   db:
     build: db/.
-    ports:
-      - "33060:3306"
     command: --default-authentication-plugin=mysql_native_password
     restart: always
