+++ src/main/java/com/napier/seprojectgroup6/App.java
@@ -1,37 +1,87 @@
 package com.napier.seprojectgroup6;
 
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
 
-import com.napier.seprojectgroup6.db.ConnectionManager;
-import com.napier.seprojectgroup6.navigation.Navigator;
-import com.napier.seprojectgroup6.reports.CitiesByDistrictReport;
 public class App {
+
     /**
-     *
-     * @param args
+     * Connection to MySQL database.
      */
-    public static void main(String[] args) {
+    private Connection con = null;
 
-        // Connect to database and instantiate Navigator
-        // if running locally i.e not on Github actions.
+    /**
+     * Connect to the MySQL database.
+     */
+    public void connect()
+    {
+        try
+        {
+            // Load Database driver
+            Class.forName("com.mysql.cj.jdbc.Driver");
+        }
+        catch (ClassNotFoundException e)
+        {
+            System.out.println("Could not load SQL driver");
+            System.exit(-1);
+        }
 
+        int retries = 10;
+        for (int i = 0; i < retries; ++i)
+        {
+            System.out.println("Connecting to database...");
+            try
+            {
+                // Wait a bit for db to start
+                Thread.sleep(30000);
+                // Connect to database
+                con = DriverManager.getConnection("jdbc:mysql://db:3306/world?useSSL=false", "root", "semgroup6");
+                System.out.println("Successfully connected");
+                break;
+            }
+            catch (SQLException sqle)
+            {
+                System.out.println("Failed to connect to database attempt " + Integer.toString(i));
+                System.out.println(sqle.getMessage());
+            }
+            catch (InterruptedException ie)
+            {
+                System.out.println("Thread interrupted? Should not happen.");
+            }
+        }
+    }
+
+    /**
+     * Disconnect from the MySQL database.
+     */
+    public void disconnect()
+    {
+        if (con != null)
+        {
+            try
+            {
+                // Close connection
+                con.close();
+            }
+            catch (Exception e)
+            {
+                System.out.println("Error closing connection to database");
+            }
+        }
+    }
+
+    public static void main(String[] args)
+    {
+        // Test Output to indicate Project is working
         System.out.println("Project Works!");
+        // Create new Application
+        App a = new App();
 
-        if(args.length < 1){
-            ConnectionManager.getInstance().connect("localhost:33060", 30000);
-        }else{
-            ConnectionManager.getInstance().connect(args[0], Integer.parseInt(args[1]));
-        }
+        // Connect to database
+        a.connect();
 
-        boolean isRunningOnDocker = args.length > 1;
-        if(!isRunningOnDocker) {
-            Navigator navigator = new Navigator();
-        }
-        else {
-
-            CitiesByDistrictReport report = new CitiesByDistrictReport();
-            report.runWithDistrict("Aichi");
-
-            ConnectionManager.getInstance().disconnect();
-        }
+        // Disconnect from database
+        a.disconnect();
     }
 }
