--- src/main/java/com/napier/seprojectgroup6/db/CapitalCity.java
@@ -1,10 +1,0 @@
-package com.napier.seprojectgroup6.db;
-
-// Field names for Report Display
-public class CapitalCity {
-
-    public String name;
-    public String country;
-    public int population;
-
-}
