--- src/main/java/com/napier/seprojectgroup6/db/City.java
@@ -1,9 +1,0 @@
-package com.napier.seprojectgroup6.db;
-
-public class City {
-    public int ID;
-    public String countryCode;
-    public String district;
-    public String name;
-    public int population;
-}
