--- src/main/java/com/napier/seprojectgroup6/db/Country.java
@@ -1,11 +1,0 @@
-package com.napier.seprojectgroup6.db;
-
-// Field names for Report Display
-public class Country {
-    public String code;
-    public String name;
-    public String continent;
-    public String region;
-    public int population;
-    public String capital;
-}
