--- src/main/java/com/napier/seprojectgroup6/db/Population.java
@@ -1,15 +1,0 @@
-package com.napier.seprojectgroup6.db;
-
-// Field names for Report Display
-public class Population {
-
-    //Name of Continent/region/country
-    public String name;
-
-    public int totalPopulation;
-
-    // Outputs from these should include a % sign
-    public double pctLivingInCities;
-    public double pctNotLivingInCities;
-
-}
