--- src/main/java/com/napier/seprojectgroup6/navigation/ReportMenu.java
@@ -1,12 +1,0 @@
-package com.napier.seprojectgroup6.navigation;
-
-public class ReportMenu {
-    String title;
-    String className;
-
-    ReportMenu(String title, String className) {
-        this.className = className;
-        this.title = title;
-    }
-
-}
