--- src/main/java/com/napier/seprojectgroup6/reports/Report.java
@@ -1,5 +1,0 @@
-package com.napier.seprojectgroup6.reports;
-
-public interface Report {
-    void run();
-}
