--- use-cases/sem-group6.puml
@@ -1,57 +1,0 @@
-@startuml
-  left to right direction
-
-  actor User as U
-
-  ' Application Interface
-  rectangle "World Reports System" as wr {
-
-      ' These Use cases all include a option to restrict output with the N Filter
-      rectangle "Area Reports" as selReport {
-        usecase "Country Reports" as countryRep
-        usecase "City Reports" as cityRep
-        usecase "Capital City Reports" as capCityRep
-      }
-
-      ' These use cases require different selections from those above
-      usecase "Population Reports" as popRep
-      usecase "Language Report" as langRep
-
-      ' User can select any report type to run
-      U --> selReport
-      U --> popRep
-      U --> langRep
-
-      ' All reports in this category have identical N filters
-      usecase "Input N Filter" as inputN
-      inputN <.. selReport : include
-      note top of inputN : Enter 0 for all records
-
-      ' These reports require different sub filters with City Reports requiring additional categories
-      usecase "Select: World, Continent, Region" as cat1
-      cat1 <.. countryRep : include
-      cat1 <.. capCityRep : include
-      usecase "Select: Country, District" as cat2
-      cat1 <.. cityRep : include
-      cat2 <.. cityRep : include
-
-      ' These reports require different subset of options
-      usecase "Living & Not living in cities in each Continent, region, country" as cat3
-      cat3 <.. popRep : include
-
-      usecase "Total Population of: World, continent, region, country, district, city" as cat4
-      cat4 <.. popRep :include
-
-      usecase "Total of people speaking: Chinese, English, Hindi, Spanish, Arabic" as cat5
-      cat5 <.. langRep : include
-  }
-
-  ' All reports require database access
-  Database world
-  wr --> world
-
-  ' All reports require a output of the results
-  usecase "Display Results" as display
-  wr --> display
-
-@enduml
